atcDataSourceTimeseriesSWMM5Output.vb: In the Open function, number of value is changed from Int16 to Int32 to hold                                        greater number swmm5_iface.vb: clean up code, remove unused commented out codes

+++ BASINS40/atcTimeseriesSWMMOutput/atcDataSourceTimeseriesSWMM5Output.vb
@@ -79,7 +79,7 @@ Public Class atcDataSourceTimeseriesSWMM5Output
                     Next
                 End If
 
-                Dim lNumValues As Int16 = pSWMM5_OutputFile.TimeStarts.GetUpperBound(0)
+                Dim lNumValues As Int32 = pSWMM5_OutputFile.TimeStarts.GetUpperBound(0)
 
                 'Set up common date array
                 Dim lDates As atcTimeseries = New atcTimeseries(Me)
